perf: do parallel when is possible
