+++ packages/multi/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@keyvhq/multicache",
+  "name": "@keyvhq/multi",
   "description": "Layered cache with any backend",
   "homepage": "https://keyv.js.org",
   "version": "1.2.7",
@@ -27,10 +27,10 @@
     "value"
   ],
   "dependencies": {
-    "@keyvhq/core": "~1.2.6",
-    "@keyvhq/keyv-sqlite": "^1.0.0"
+    "@keyvhq/core": "~1.2.6"
   },
   "devDependencies": {
+    "@keyvhq/keyv-sqlite": "latest",
     "ava": "latest",
     "delay": "latest",
     "nyc": "latest"
